requirements.txt
backports.shutil-get-terminal-size==1.0.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.3.0
enum34==1.1.6
Flask==1.0.2
gunicorn==19.8.1
idna==2.6
ipdb==0.11
ipython==5.7.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycrypto==2.6.1
Pygments==2.2.0
requests==2.18.4
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
